requirements.txt
langchain
openai
faiss-cpu
tiktoken
python-dotenv


